requirements.txt
torch
transformers
wandb
pillow
accelerate